bare bones test
In order to get out of the conceptual and into the concrete, I wanted to make a functional prototype. This is just a bare bones manifest extension, without many of the features that I want the final to have
currently it successfully blocks two websites (reddit and twitch) with the help of the antiReddit.json file that establishes the ruleset. 

however this means that 
1. there isn't an options page to change the website you want to block
1b. there isn't an oninstalled prompt to set up a website to block
2. the blocking doesn't toggle on or off, it is permanently on
2b. no time is allowed to see the site before toggling it off
2c. there's no timer to turn it back on when it's toggled off
2d. there's no indicator on the button whether the ruleset is active/inactive and how much time is left

So before I figure out how to change the ruleset website I'm going to see if I can get the ruleset to toggle on/off