Create serviceTester
this is the background service worker file, saved as service tester until we're done with it. 

currently the functionalities in it are just a test block of code to see if an onvisited function works (webNav onCompleted)
and an onInstalled function to set the badge to be active. Will eventually replace oninstalled with an onvisited function that has an alarm timer to turn it off