Update serviceTester
haven't figured out how to do notifications yet. I've got some in there, and code seems to be working, but no notifications are popping up. Not sure what they're supposed to look like yet though.

I also rearranged where to put the activateBlocker function call.
Before I had it when the alarm was being made, which immediately blocked the website and prevented subsequent openings to the site or in-site navigation. Which was annoying for the time period that you're supposed to be allowed on the site.
By moving that to the alarm-trigger section the blocker is put up and then you're booted off all your tabs, with a notification that your time on the site has run out. Makes more sense.