Update serviceTester
WHEW it was hard to figure out notifications.

had to spend a long time figuring out that I had actually muted all chrome notifications in windows 10. lol shouts out to the people at chrome support and the guys on the chromium forums.
but finally got the notifications working and populating.
Had to inject an if statement to make sure the onCompleted function didn't create duplicates anymore.

but I did it! Now next I can put notification creators and clearers at all the relevant points in the ruleset deac/reactivation cycle.

Also created a second badge color for when your time on the site begins to tick down before it's actually blocked.