Update serviceTester
hit a block trying to create functions to convert a list of url strings into an object that the listener urlFilter property could read. 

per broscious's suggestion, instead of trying to make the listeners en masse, I've changed the code to make a different listener for each restricted website instead.
I broke the oncompleted logic into its own function to be called by the newly created listeners, and now I can create listeners for any amount of sites that I want. later they'll be created onbuttonclick when user updates their options, but for now they're just created on installation to test.

also laid out what the rest of the onbuttonclick fn needs to do - create a storage dict of all the sites with rules, and create a dynamic ruleset for the site.

but now that this currently works I need to test if I can delete listeners that I've created.