Update and rename serviceTester to serviceWorker
added command to remove a site. successfully fully updates storage object and removes listener for the site.
Though it will wait until blocker is deactivated if you remove a site in the middle of an activation.

next is to do the front end and get a functional options page, and to implement buttons to contact the service worker.