Update optionsPage.js
got the delete buttons to work. sort of.
they're laggy, they're supposed to refresh the list of current websites that are restricted, but that's spotty.
Only the top delete button works. Seems like the listeners aren't working correctly for the other links. But when one is deleted the next starts working.

Also haven't fixed sync glitch