Update serviceWorker.js
I realized the reason that the links were overwriting was because the code wasn't waiting for the storage.sync.get ter to finish before progressing with the site preferences update.
Not sure if this is a dumb way to do this, but I couldn't get .then or await logic to work, so I stuffed the rest of the code in a settimeout so that it would have to wait for the event stack to clear before proceeding. Now when I terminate the extension and open it back up (as if on inactivity) the site settings persist.