Update serviceWorker.js
Two thoughts on how to maintain the listeners so that they are active whenever the user needs them - 
Either put them before every chance that the user could be navigating to a website, meaning that before oncompleted you'd have to put code onnavigated to add the listeners.
Seems like a lot.
This option is to refresh the listeners every time the state of the machine changes. On idle and on activate the whole cache of websites will be deleted and readded so that the listeners are fresh. 
Hopefully this works. It hasn't broken the code, but about to go through testing.