Update optionsPage.js
well while we're doing a big 1.1 update I wanted to add a feature for better user visibility.
Before the timer slider was just stuck on the default setting of 1h, 23hrs. But now I've injected some logic into the populateSettings command so that every time it pulls the stored links, it'll also pull the stored time settings and set the sliders to that value instead of the default.
It's a nice little upgrade.