Update serviceWorker.js
Had a hard time constructing promises to strip down the entire cache of urls and get them back up to refresh the listeners.
Was also a little more involved than this.
Thought of a slightly more elegant solution, though maybe heavy handed?
Every time the user goes from inactivity or locked to an active state, the extension will simply delete all the listeners, recreate them all, update the cache, and then re-store the new cache to ensure that the new listener functions match.
Brief (1 hr) testing (that was sufficient to show when the program was broken before) has shown this to work so far. Need to give it a more long term test though.