Update optionsPage.html
added more timer buttons for /1hr and /3hr for better user functionality. If the user inputs more time allowed than their per button allottment, the function catches it and minimizes the amount of time blocked to 15mins.

If anna helps with QA testing need to add her to the thank you list in here.