Update serviceWorker.js
successfully added in a check to make sure that we didn't get alternate listener hits on the oncompleted to the website

but I started getting issues with the countdown timer after that. Some flickering when the second timer activated because the second was past 0 and not deleting. So I had to change some logic to make it work right. i think i had an extra clear interval in the initiate badge portion I didn't need. 
In any case it seems it's all working now!