Update serviceWorker.js
I was able to create a better method of creating and removing the intervals by storing all the intervals we're creating in a set inside a storage object, and then clearing all the intervals in our set every time one is finished.
It seems to have fixed our issues with the flickering timer and also the issue where intervals have been sneaking through the clear interval stage causing us to go into negative counting for infinity. Even though we're creating tons of intervals they all get saved in the object so they're all being removed effectively.

The only issue I have with this fix is that I realize now that we're still creating intervals for every time we successfully access one of our restricted websites throughout the warning period.
So in a similar way to how I added a variable to validate the first settimeout for inactivity, next I'm going to validate whether or not we've already triggered on our oncompleted so that we don't play through all the logic every time we visit a site during the warning period.
I don't think it's 100% necessary, but it should technically be more performant, since I stress tested and was able to create a few hundred timers in just 60 seconds of warning period.