Update serviceWorker.js
added a variable to note whether we've visited one of our restricted website, so that we don't repeat the oncompleted logic multiple times during our warning period.
visited changes from false to true on our first completed visit to a restricted website, and then changes back to false when the alarm goes off to turn on our blocker.

Although this didn't eliminate 100% of the extra settimeout calls, it did eliminate the majority of them (increased from 3 to 5 to 7 to 16 rather than 3 to 200 or so varying wildly on amount of clicks)
However, with our other implemented logic including the timeouts in our set, this is no longer an issue!
So everything appears to be working perfectly right now! I've commented out the console log for our leak testing, and it's time to ship out version 2.1 and do another month of use testing.